feat : nudge position 기능 추가

+++ src/main/java/com/team6/onandthefarmsnsservice/entity/FeedImageProduct.java
@@ -29,7 +29,7 @@ public class FeedImageProduct {
 
     private Long productId;
 
-    private Double positionX;
+    private Integer positionX;
 
-    private Double positionY;
+    private Integer positionY;
 }
